requirements.txt
robotpy
robotpy-commands-v2
robotpy-halsim-gui
robotpy-rev
wpilib==2026.2.1
certifi
